+++ src/test/java/org/folio/rest/api/BoundWithStorageTest.java
@@ -12,10 +12,12 @@ import java.net.HttpURLConnection;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import junitparams.JUnitParamsRunner;
 import lombok.SneakyThrows;
+import org.folio.okapi.common.XOkapiHeaders;
 import org.folio.rest.support.IndividualResource;
 import org.folio.rest.support.Response;
 import org.folio.rest.support.ResponseHandler;
@@ -418,10 +420,11 @@ public class BoundWithStorageTest extends TestBaseWithInventoryUtil {
   }
 
   private IndividualResource createHoldingsRecord(UUID instanceId) {
-    return holdingsClient.create(
-      new HoldingRequestBuilder()
-        .forInstance(instanceId)
-        .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID));
+    HoldingRequestBuilder holdingRequestBuilder = new HoldingRequestBuilder()
+      .forInstance(instanceId)
+      .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID);
+    return holdingsClient.create(holdingRequestBuilder.create(), TENANT_ID,
+      Map.of(XOkapiHeaders.URL, mockServer.baseUrl()));
   }
 
   private IndividualResource createItem(UUID holdingsRecordId) {