+++ src/test/java/org/folio/rest/api/ItemEffectiveCallNumberComponentsTest.java
@@ -6,6 +6,7 @@ import static org.folio.rest.support.matchers.ItemMatchers.hasCallNumber;
 import static org.folio.rest.support.matchers.ItemMatchers.hasPrefix;
 import static org.folio.rest.support.matchers.ItemMatchers.hasSuffix;
 import static org.folio.rest.support.matchers.ItemMatchers.hasTypeId;
+import static org.folio.utility.RestUtility.TENANT_ID;
 import static org.hamcrest.CoreMatchers.allOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -14,12 +15,14 @@ import static org.junit.Assert.assertNotNull;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import java.net.HttpURLConnection;
+import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
 import junitparams.JUnitParamsRunner;
 import junitparams.Parameters;
 import junitparams.naming.TestCaseName;
 import org.apache.commons.lang3.StringUtils;
+import org.folio.okapi.common.XOkapiHeaders;
 import org.folio.rest.api.testdata.ItemEffectiveCallNumberComponentsTestData;
 import org.folio.rest.api.testdata.ItemEffectiveCallNumberComponentsTestData.CallNumberComponentPropertyNames;
 import org.folio.rest.support.IndividualResource;
@@ -275,13 +278,14 @@ public class ItemEffectiveCallNumberComponentsTest extends TestBaseWithInventory
 
     IndividualResource instance = instancesClient.create(instance(UUID.randomUUID()));
 
-    IndividualResource holdings = holdingsClient.create(new HoldingRequestBuilder()
+    JsonObject holdingToCreate = new HoldingRequestBuilder()
       .withId(UUID.randomUUID())
       .forInstance(instance.getId())
       .withPermanentLocation(MAIN_LIBRARY_LOCATION_ID)
       .create()
-      .put(propertyName, propertyValue)
-    );
+      .put(propertyName, propertyValue);
+    IndividualResource holdings =
+      holdingsClient.create(holdingToCreate, TENANT_ID, Map.of(XOkapiHeaders.URL, mockServer.baseUrl()));
     assertThat(holdings.getJson().getString(propertyName), is(propertyValue));
 
     return holdings;